test: adding hotjar
hotjar wrap for usability testing
